feat: Add companyName field to user registration
This commit adds the `companyName` field to the user registration process. When registering a user, the `companyName` can now be provided in the request body. If the `companyName` is provided, it will be used as the user's name in the `profileData` object. If the `companyName` is not provided, the user's first name will be used instead.

Refactor the `registerUser` function in the `user.controller.js` file to include the `companyName` field in the destructured request body. Update the `profileData` object to use the `companyName` if it is provided, otherwise use the user's first name.

This change improves the user registration process by allowing users to provide their company name during registration.