refactor: Update addRecommendation function to include recommendedBy field
Update the addRecommendation function in the individualUser controller to include the recommendedBy field in the request body. This field specifies the user who recommended the job. If the recommendedBy field is provided, it will be used as the fromUserId instead of retrieving it from the token. This change allows for more flexibility in recommending jobs and improves the accuracy of the recommendation process.

Note: This commit message follows the established conventions in the repository.